+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All Notable changes to `gent/services-opening-hours` package.
 
-## [Unreleased]
+## [0.2.0]
 
 ### Added
 
@@ -35,5 +35,6 @@ hours for a Channel.
 * DMOH-17: Added the OpeningHoursPeriod & OpeningHoursPeriodHTML methods to the
   ChannelService.
 
+[0.2.0]: https://github.com/StadGent/php_package_services-opening-hours/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/StadGent/php_package_services-opening-hours/releases/tag/0.1.0
 [Unreleased]: https://github.com/StadGent/php_package_services-opening-hours/compare/master...develop